Angular Project - Search for Products by Category - Code - Enhance ProductList Completed
Section 12 - Search for Products by Category: 78 Completed
Angular Project - Search for Products by Category - Code - Enhance ProductList Completed
Added the routes to the ProductListComponent. Then checking if there is a value actually being passed for the currentCategoryId or not, then fetching that value, converting it to a number and then obtaining the list of products for that particular Category ID by subscribing to the productService() which is mapped to the productService() in the product.service.ts file.