Angular Project - Dynamic Search Component - Create Component Completed
Section 13 - Search for Products by Category (Dynamic Search Component): 85 Completed
Angular Project - Dynamic Search Component - Create Component
ProductCategoryMenuComponent is created in which we define an Array of ProductCategory and then inject the Component Service. Further, we define the listProductCategories() method to invoke the service and assign the data to the Property.