Angular Project - Search for Products by Keyword - Update ProductService API
Section 14: Search for Products by Keyword - 97 Completed
Angular Project - Search for Products by Keyword - Update ProductService API
Created the searchProducts() method to search for products based on the keyword provided by the user.
Refactored the code to create a single method for the class which would search based on the custom URL generated based on the input received.